pyproject.toml
    [project]
name = "py-roguelike-tutorial"
version = "0.1.0"
description = ""
authors = [
    {name = "Mirco Kraenz",email = "software@kraenz.eu"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "tcod>=19.0.0",
]

[tool.poetry]
packages = [{include = "py_roguelike_tutorial", from = "src"}]

[too.pyright]
typeCheckingMode = "strict"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "cx-freeze>=8.3.0",
]
